KLC/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-js-asset==2.0.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.4.2
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.5.0
pip==23.1.2
pycparser==2.21
pytz==2023.3
requests==2.30.0
setuptools==65.5.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
wheel==0.38.4
wrapt==1.15.0
